fix(typing): resolve invalid-assignment and invalid-return-type
Fix the 136 ty hits of these two rules and promote them to errors.

Fixes at the source:
- `safe_divide` gets overloads (scalar/scalar -> float, any array ->
  FloatArray); its 81 call sites no longer see `float | FloatArray`.
- `AssetPanel`/`AssetPanelView.__getitem__` get overloads: a `str` key
  returns the field array, any other key an `AssetPanelView`.
- `input_to_array` returns `AnyArray` (it also converts group labels).
- `BaseCovariance.mahalanobis` returns `float | FloatArray` (float for
  1-D input).
- `_validate_estimators` / `_validate_factors` return tuples, as
  `zip(*...)` produces.
- `_compute_top_indices` returns integer index arrays.
- `binary_train_test_sets` / `index_train_test_` are float arrays
  (built with `np.zeros`).
- `is_positive_definite` / `compute_optimal_n_clusters` return builtin
  `bool` / `int`.
- A reused parameter name in `_combine_scores` gets its own local, and
  `indexer` in `FactorModel` is annotated.

The remaining 79 hits are suppressed inline with a reason: fitted
attributes reset to None before refit, dataclass `None` defaults filled
in `__post_init__`, cvxpy/numpy stub imprecision, and invariants ty
cannot follow.

Refs #365

+++ src/skfolio/attribution/_model/_attribution.py
@@ -29,7 +29,7 @@ from skfolio.attribution._utils import (
     _format_decimal,
     _format_percent,
 )
-from skfolio.typing import AnyArray, FloatArray
+from skfolio.typing import AnyArray, FloatArray, IntArray
 from skfolio.utils.figure import format_plot_label
 
 __all__ = ["Attribution"]
@@ -1008,7 +1008,7 @@ def _slice_window(
 
 def _compute_top_indices(
     sort_values: FloatArray, top_n: int | None
-) -> tuple[FloatArray, FloatArray | None]:
+) -> tuple[IntArray, IntArray | None]:
     """Compute indices for top_n selection and "Other" aggregation.
 
     Parameters
